+++ examples/java/Sdd.java
@@ -1,4 +1,4 @@
-//> using dep "io.github.riccardomerolla:llm4zio-java:3.16.0"
+//> using dep "io.github.riccardomerolla:llm4zio-java:3.17.0"
 //> using scala "3.8.3"
 //> using jvm 21